fix: address PR #13 review findings
- [BLOCKER] SubscriberRow.RenewalMonth/RenewalDay and LawChangeRow.RetryCount
  were long/long? to match SQLite's INTEGER affinity; Npgsql maps Postgres
  INTEGER to Int32, and Dapper's record-constructor materialization requires an
  exact type match (unlike class property assignment, which coerces). Fixed to
  int/int?. Verified empirically: the original long types throw
  InvalidOperationException against real Postgres; the fixed int types don't.

- Added prevent_destroy to the Postgres Flexible Server, matching the AKS
  cluster/Key Vault/ACR guards already in resources.tf — it was the one
  stateful resource left unprotected from a forced-replacement plan.

- Added resource requests/limits and a baseline securityContext
  (runAsNonRoot, allowPrivilegeEscalation: false, capabilities drop ALL) to
  all seven Helm workload templates. Sized per workload (crawler gets the
  most, portal the least) in values.yaml.

- Made the non-root securityContext actually work end-to-end rather than just
  declaring it:
  - portal: swapped to nginxinc/nginx-unprivileged (port 8080, can't bind 80
    as non-root); verified the built image serves correctly as the `nginx`
    user.
  - crawler: added a non-root user in the Dockerfile, --no-sandbox on
    Chromium's launch (its own sandbox needs capabilities the pod
    deliberately drops), and a /dev/shm emptyDir (K8s's default is too small
    for Chromium). Verified Chromium actually launches under this
    configuration, on both arm64 and the amd64 architecture GitHub Actions
    actually builds on.
  - Found and fixed two version-drift bugs surfaced by that verification:
    portal's package-lock.json was missing Linux-only optional deps (only
    ever regenerated on macOS), and crawler's Dockerfile pinned a Playwright
    image tag that didn't match package.json's floating playwright range.
    Pinned playwright to an exact version and regenerated both lockfiles
    against linux/amd64 to match CI.

- design.md said the app namespace was `nomadrules`; every Terraform/Helm
  file actually uses `nomadrules-services`. Fixed the doc to match what was
  built.

+++ src/email-service/NomadRules.EmailDelivery/Delivery/Models.cs
@@ -16,8 +16,9 @@ public class DeliveryOptions
 
 // A subscriber and the renewal month (1-12) + optional day (1-31) for one category. Month is null when unset
 // for that category; day is null when only a month was entered (anchor falls back to the 1st).
-// long? to match SQLite INTEGER (Dapper requires an exact type match).
-public record SubscriberRow(string Id, string Email, string State, long? RenewalMonth, long? RenewalDay);
+// int? to match Postgres INTEGER (Dapper's record/constructor materialization requires an exact type match —
+// unlike class property assignment, which tolerates numeric coercion).
+public record SubscriberRow(string Id, string Email, string State, int? RenewalMonth, int? RenewalDay);
 
 // Just enough to address a subscriber, for state-matched digest/urgent delivery — no renewal fields, so
 // insurance-specific columns can't be misread as a generic renewal value.
